backend/requirements.txt
# Core Framework
Django>=5.1,<5.2

# API Framework
djangorestframework>=3.15.2

# Authentication
djangorestframework-simplejwt>=5.5.0

# CORS Management
django-cors-headers>=4.7.0

# Supporting dependencies (auto-installed, but good to pin)
PyJWT>=2.10.0
cryptography>=42.0.0